requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
asttokens==3.0.0
attrs==24.3.0
blessed==1.20.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
decorator==5.1.1
diskcache==5.6.3
executing==2.1.0
gpustat==1.1.1
guidance==0.2.0
guidance-stitch==0.1.0
idna==3.10
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jinxed==1.3.0
jupyterlab_widgets==3.0.13
llama_cpp_python==0.3.6
llguidance==0.5.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
numpy==2.2.2
nvidia-ml-py==12.560.30
ordered-set==4.1.0
parso==0.8.4
pathvalidate==3.2.3
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
stack-data==0.6.3
tiktoken==0.8.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
